requirements.txt
Django==1.9.1
ipython==4.0.0
ipdb==0.8.1
pillow==3.1.0
django-extensions==1.6.1
django-debug-toolbar==1.4
django-debug-toolbar-template-profiler==1.0.1
django-widget-tweaks==1.4.1
django-storages==1.1.8
boto==2.39.0
psycopg2==2.6.1